# frontend/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CarDian",
        "version": "0.1.0"
    },
    "paths": {
        "/v1/users/profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get current user profile",
                "operationId": "get_my_profile_v1_users_profile_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponseSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/v1/users/email": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Change user's email",
                "operationId": "change_user_email_v1_users_email_patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailChangeRequestSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/v1/users/fullname": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Change user's full name",
                "operationId": "change_user_name_v1_users_fullname_patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FullNameChangeRequestSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/v1/users/password": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Change user's password",
                "operationId": "change_password_v1_users_password_patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordChangeRequestSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/v1/users/me": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deactivate users profile",
                "operationId": "deactivate_profile_v1_users_me_delete",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteUserRequestSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteUserResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
        "/v1/users/payments/": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get current user payments",
                "operationId": "get_user_payments_v1_users_payments__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentSequenceResponseSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Complete payment after processing (credit the tokens)",
                "operationId": "complete_payment_v1_users_payments__patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentRequestSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/v1/users/payments/{payment_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get current user payment by id",
                "operationId": "get_payment_by_id_v1_users_payments__payment_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "payment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Payment Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DeleteUserRequestSchema": {
                "properties": {
                    "current_password": {
                        "type": "string",
                        "title": "Current Password"
                    }
                },
                "type": "object",
                "required": [
                    "current_password"
                ],
                "title": "DeleteUserRequestSchema"
            },
            "DeleteUserResponseSchema": {
                "properties": {
                    "is_deleted": {
                        "type": "boolean",
                        "title": "Is Deleted"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Deleted At"
                    }
                },
                "type": "object",
                "required": [
                    "is_deleted",
                    "deleted_at"
                ],
                "title": "DeleteUserResponseSchema"
            },
...
            "EmailChangeRequestSchema": {
                "properties": {
                    "new_email": {
                        "type": "string",
                        "format": "email",
                        "title": "New Email"
                    }
                },
                "type": "object",
                "required": [
                    "new_email"
                ],
                "title": "EmailChangeRequestSchema"
            },
...
            "FullNameChangeRequestSchema": {
                "properties": {
                    "new_full_name": {
                        "type": "string",
                        "title": "New Full Name"
                    }
                },
                "type": "object",
                "required": [
                    "new_full_name"
                ],
                "title": "FullNameChangeRequestSchema"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PasswordChangeRequestSchema": {
                "properties": {
                    "old_password": {
                        "type": "string",
                        "title": "Old Password"
                    },
                    "new_password": {
                        "type": "string",
                        "title": "New Password"
                    }
                },
                "type": "object",
                "required": [
                    "old_password",
                    "new_password"
                ],
                "title": "PasswordChangeRequestSchema"
            },
            "PaymentRequestSchema": {
                "properties": {
                    "rate": {
                        "$ref": "#/components/schemas/RateTypes"
                    },
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    }
                },
                "type": "object",
                "required": [
                    "rate",
                    "user_id"
                ],
                "title": "PaymentRequestSchema"
            },
            "PaymentResponseSchema": {
                "properties": {
                    "rate": {
                        "$ref": "#/components/schemas/RateTypes"
                    },
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "status": {
                        "$ref": "#/components/schemas/PaymentStatus"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "rate",
                    "user_id",
                    "id",
                    "status",
                    "created_at"
                ],
                "title": "PaymentResponseSchema"
            },
            "PaymentSequenceResponseSchema": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/PaymentResponseSchema"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "total_items_qty": {
                        "type": "integer",
                        "title": "Total Items Qty"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "total_items_qty"
                ],
                "title": "PaymentSequenceResponseSchema"
            },
            "PaymentStatus": {
                "type": "string",
                "enum": [
                    "success",
                    "failed",
                    "pending"
                ],
                "title": "PaymentStatus"
            },
...
            "RateTypes": {
                "type": "string",
                "enum": [
                    "starter",
                    "pro"
                ],
                "title": "RateTypes"
            },
...
            "UserResponseSchema": {
                "properties": {
                    "full_name": {
                        "type": "string",
                        "title": "Full Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "role": {
                        "type": "string",
                        "title": "Role"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "current_balance": {
                        "type": "integer",
                        "title": "Current Balance"
                    }
                },
                "type": "object",
                "required": [
                    "full_name",
                    "email",
                    "id",
                    "role",
                    "created_at",
                    "current_balance"
                ],
                "title": "UserResponseSchema"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/v1/auth/login"
                    }
                }
            }
        }
    }
}